+++ src/webapp/tests/e2e/specs/sqlite-carbon-2.spec.js
@@ -4,10 +4,9 @@ describe('Configuration', () => {
     cy.route('/api/v1/repos/models').as('modelsApi')
     cy.route('/api/v1/reports').as('reportsApi')
 
-    cy.route(
-      'POST',
-      '/api/v1/sql/get/model-carbon-intensity-sqlite/carbon/region'
-    ).as('getChartApi')
+    cy.route('POST', '/api/v1/sql/get/model-carbon-intensity/carbon/region').as(
+      'getChartApi'
+    )
     cy.route('POST', '/api/v1/dashboards/dashboard/save').as('saveDashboard')
     cy.route('POST', '/api/v1/dashboards/dashboard/report/add').as('addReport')
     cy.route('POST', '/api/v1/reports/save').as('saveReportsApi')
@@ -20,13 +19,13 @@ describe('Configuration', () => {
     cy.get('.tag-running-pipelines').should('have.length', 0)
     cy.visit('/model')
     cy.wait('@modelsApi')
-    cy.get(
-      '[data-test-id="model-carbon-intensity-sqlite-carbon-model-card"]'
-    ).within(() => {
-      cy.get('.button')
-        .contains('Analyze')
-        .click()
-    })
+    cy.get('[data-test-id="model-carbon-intensity-carbon-model-card"]').within(
+      () => {
+        cy.get('.button')
+          .contains('Analyze')
+          .click()
+      }
+    )
     cy.wait('@reportsApi')
     cy.get('[data-test-id="column-name"]').click()
     cy.get('[data-test-id="aggregate-count"]').click()